Create basic Docker build
This also includes a starter Makefile which makes running the container
easier. We should look at updating it as a separate story later.

Also, this container needs to run as root or the app can't access the
H2 DB. Not sure why this happens, but will fix it only if the H2 DB
is actually needed.